+++ packages/fury-adapter-oas3-parser/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Object properties in a Schema Object are now supported.
 - Referencing (`$ref`) a response object is now supported.
 - Array items in a Schema Object are now supported.
+- Object required properties in a Schema Object are now supported.
 
 ## 0.4.1 (28-01-19)